v2o.4.1

+++ src/main/java/nc/multiblock/gui/GuiTurbineController.java
@@ -61,8 +61,8 @@ public class GuiTurbineController extends GuiMultiblock<Turbine, ITurbineControl
 			inputRate = Lang.localise("gui.nc.container.turbine_controller.power_bonus") + " " + NCMath.decimalPlaces(100D * multiblock.powerBonus, 1) + "%";
 		}
 		else {
-			double rateRatio = (double) multiblock.recipeInputRate / (double) multiblock.getLogic().getMaxRecipeRateMultiplier();
-			double rateRatioFP = (double) multiblock.recipeInputRateFP / (double) multiblock.getLogic().getMaxRecipeRateMultiplier();
+			double rateRatio = multiblock.recipeInputRate / (double) multiblock.getLogic().getMaxRecipeRateMultiplier();
+			double rateRatioFP = multiblock.recipeInputRateFP / multiblock.getLogic().getMaxRecipeRateMultiplier();
 			inputRate = Lang.localise("gui.nc.container.turbine_controller.fluid_rate") + " " + UnitHelper.prefix(multiblock.recipeInputRateFP, 6, "B/t", -1) + " [" + Math.round(100D * rateRatioFP) + (rateRatio > 1D ? "%] [!]" : "%]");
 			inputRateWidth = inputRateWidth - width(inputRate) > 1 ? width(inputRate) : Math.max(inputRateWidth, width(inputRate));
 		}
